requirements.txt
# --- Core Frameworks ---
streamlit>=1.35.0
pandas>=2.0.0

# --- LangChain Ecosystem ---
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
langchain-huggingface>=0.0.3
langchain-experimental>=0.0.60

# --- Vector Database & Embeddings ---
faiss-cpu>=1.8.0
sentence-transformers>=2.2.0

# --- Environment & Utilities ---
python-dotenv>=1.0.0
numpy>=1.24.0

statsmodels>=0.14.0
